pyproject.toml
[tool.poetry]
name = "zood"
version = "0.0.10"
description = "web page documentation & comment generation documentation"
license = "MIT"
authors = ["kamilu <luzhixing12345@163.com>"]
readme = "README.md"
repository = "https://github.com/luzhixing12345/zood"
documentation = "https://luzhixing12345.github.io/zood/"


[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
zood = 'zood.main:main'
